src/crawler/requirements.txt
boto3==1.26.13
botocore==1.29.13
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
jmespath==1.0.1
numpy==1.23.4
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
